+++ src/host/coreaudio/macos/device.rs
@@ -447,7 +447,7 @@ impl Device {
         // We now check if the returned uid is non-null before use.
         if !uid.is_null() {
             let uid_string = unsafe { CFString::wrap_under_create_rule(uid).to_string() };
-            Ok(DeviceId::CoreAudio(uid_string))
+            Ok(DeviceId(crate::platform::HostId::CoreAudio, uid_string))
         } else {
             Err(DeviceIdError::BackendSpecific {
                 err: BackendSpecificError {