+++ src/host/mod.rs
@@ -37,6 +37,18 @@ pub(crate) mod webaudio;
 
 #[cfg(feature = "custom")]
 pub(crate) mod custom;
+#[cfg(not(any(
+    windows,
+    target_os = "linux",
+    target_os = "dragonfly",
+    target_os = "freebsd",
+    target_os = "netbsd",
+    target_os = "macos",
+    target_os = "ios",
+    target_os = "emscripten",
+    target_os = "android",
+    all(target_arch = "wasm32", feature = "wasm-bindgen"),
+)))]
 pub(crate) mod null;
 
 /// Compile-time assertion that a type implements Send.