+++ src/host/wasapi/device.rs
@@ -446,7 +446,7 @@ impl Device {
         unsafe {
             match self.device.GetId() {
                 Ok(pwstr) => match pwstr.to_string() {
-                    Ok(id_str) => Ok(DeviceId::Wasapi(id_str)),
+                    Ok(id_str) => Ok(DeviceId(crate::platform::HostId::Wasapi, id_str)),
                     Err(e) => Err(DeviceIdError::BackendSpecific {
                         err: BackendSpecificError {
                             description: format!("Failed to convert device ID to string: {}", e),