+++ src/host/webaudio/mod.rs
@@ -95,7 +95,10 @@ impl Device {
     }
 
     fn id(&self) -> Result<DeviceId, DeviceIdError> {
-        Ok(DeviceId::WebAudio("default".to_string()))
+        Ok(DeviceId(
+            crate::platform::HostId::WebAudio,
+            "default".to_string(),
+        ))
     }
 
     fn supported_input_configs(